+++ core/base/src/main/java/org/eclipse/dataspaceconnector/core/CoreServicesExtension.java
@@ -31,9 +31,12 @@ import org.eclipse.dataspaceconnector.spi.EdcSetting;
 import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgentService;
 import org.eclipse.dataspaceconnector.spi.command.CommandHandlerRegistry;
 import org.eclipse.dataspaceconnector.spi.message.RemoteMessageDispatcherRegistry;
+import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.policy.PolicyEngine;
 import org.eclipse.dataspaceconnector.spi.policy.RuleBindingRegistry;
+import org.eclipse.dataspaceconnector.spi.security.CertificateResolver;
 import org.eclipse.dataspaceconnector.spi.security.PrivateKeyResolver;
+import org.eclipse.dataspaceconnector.spi.security.Vault;
 import org.eclipse.dataspaceconnector.spi.system.BaseExtension;
 import org.eclipse.dataspaceconnector.spi.system.ExecutorInstrumentation;
 import org.eclipse.dataspaceconnector.spi.system.Hostname;
@@ -43,6 +46,8 @@ import org.eclipse.dataspaceconnector.spi.system.Provides;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.eclipse.dataspaceconnector.spi.system.health.HealthCheckService;
+import org.eclipse.dataspaceconnector.spi.telemetry.Telemetry;
+import org.eclipse.dataspaceconnector.spi.types.TypeManager;
 
 import java.security.PrivateKey;
 import java.time.Duration;
@@ -51,7 +56,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 
 import static java.util.Optional.ofNullable;
-
 
 @BaseExtension
 @Provides({
@@ -62,6 +66,12 @@ import static java.util.Optional.ofNullable;
         PolicyEngine.class,
         RemoteMessageDispatcherRegistry.class,
         RuleBindingRegistry.class,
+        Monitor.class,
+        TypeManager.class,
+        Vault.class,
+        PrivateKeyResolver.class,
+        CertificateResolver.class,
+        Telemetry.class
 })
 public class CoreServicesExtension implements ServiceExtension {
 
@@ -109,7 +119,8 @@ public class CoreServicesExtension implements ServiceExtension {
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        registerParser(context);
+        privateKeyResolver(context).addParser(PrivateKey.class, new DefaultPrivateKeyParseFunction());
+
         var config = getHealthCheckConfig(context);
 
         // health check service
@@ -177,7 +188,6 @@ public class CoreServicesExtension implements ServiceExtension {
 
     @Provider
     public RuleBindingRegistry ruleBindingRegistry() {
-
         return ruleBindingRegistry;
     }
 
@@ -200,6 +210,36 @@ public class CoreServicesExtension implements ServiceExtension {
         return client;
     }
 
+    @Provider
+    public Monitor monitor(ServiceExtensionContext context) {
+        return context.getMonitor();
+    }
+
+    @Provider
+    public TypeManager typeManager(ServiceExtensionContext context) {
+        return context.getTypeManager();
+    }
+
+    @Provider
+    public Telemetry telemetry(ServiceExtensionContext context) {
+        return context.getTelemetry();
+    }
+
+    @Provider
+    public Vault vault(ServiceExtensionContext context) {
+        return context.getService(Vault.class);
+    }
+
+    @Provider
+    public PrivateKeyResolver privateKeyResolver(ServiceExtensionContext context) {
+        return context.getService(PrivateKeyResolver.class);
+    }
+
+    @Provider
+    public CertificateResolver certificateResolver(ServiceExtensionContext context) {
+        return context.getService(CertificateResolver.class);
+    }
+
     private HealthCheckServiceConfiguration getHealthCheckConfig(ServiceExtensionContext context) {
 
         return HealthCheckServiceConfiguration.Builder.newInstance()
@@ -211,9 +251,5 @@ public class CoreServicesExtension implements ServiceExtension {
                 .build();
     }
 
-    private void registerParser(ServiceExtensionContext context) {
-        var resolver = context.getService(PrivateKeyResolver.class);
-        resolver.addParser(PrivateKey.class, new DefaultPrivateKeyParseFunction());
-    }
-
 }
+