+++ CHANGELOG.md
@@ -65,6 +65,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Always use configured IDS API path in IDS webhook address (#1249)
 * Fix Azure storage transfer (#1245)
 * Throw exception if `IdentityProviderKeyResolver` cannot get keys at startup (#1266)
+* Make all the services injectable (#1285)
 
 ## [milestone-3] - 2022-04-08